% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{cAttr_BF}
\alias{cAttr_BF}
\alias{cAttr}
\title{cAttr}
\usage{
cAttr_BF(eav)

cAttr(eav)
}
\arguments{
\item{eav}{effective attribute value (integer, vector)}
}
\value{
A CSFB object with 4 vectors:
`Critical`, `Success`, `Fail`, `Botch`.
}
\description{
Determine the probabilities for the outcomes of an
attribute check.
}
\section{Functions}{
\itemize{
\item \code{cAttr_BF()}: Uses a brute force algorithm. It is
intended for analysis purposes rather than practical use.

}}
\examples{
cAttr(12)
}
\seealso{
\link{cCombat}, \link{cSkill}, \link{CSFB}
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
