% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attack.R
\name{cCombat}
\alias{cCombat}
\title{cAttack
The CSFB probabilities for an attack, parry, or dodge rolls.}
\usage{
cCombat(eav, bl = 20L)
}
\arguments{
\item{eav}{the effective attribute to roll against.}

\item{bl}{The botch level, typically only a 20 is a botch, but some
weapons or maneouvers change that.}
}
\value{
A list with 4 vectors: `Critical`, `Success`, `Fail`, `Botch`
(S3 class `CSFB`).
}
\description{
cAttack
The CSFB probabilities for an attack, parry, or dodge rolls.
}
\examples{
cCombat(12)
}
