% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convolveDice}
\alias{convolveDice}
\title{convolveDice}
\usage{
convolveDice(x, y)
}
\arguments{
\item{x, y}{The two vectors to convolve.}
}
\value{
Vector of length `length(x) + length(y)`.
}
\description{
Convolves two distributions of [n]d[M].
}
