% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{crit3d20}
\alias{crit3d20}
\title{in a 3d20 which are defined by at least two instances
of a 1 in that roll.}
\usage{
crit3d20(eav)
}
\description{
in a 3d20 which are defined by at least two instances
of a 1 in that roll.
}
