% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{cSkill}
\alias{cSkill}
\title{cSkill}
\usage{
cSkill(eav, skill)
}
\arguments{
\item{eav}{A vector with 3 places, each being an effective
attribute value.}

\item{skill}{The skill value}
}
\value{
A CSFB object with 4 vectors:
`Critical`, `Success`, `Fail`, `Botch`.
}
\description{
Return the aggregated probability for a critical success,
a regular success, a fail, and a botch.
}
\examples{
cSkill(c(11, 12, 14), 7)
}
\seealso{
\link{cAttr}, \link{cCombat}, \link{CSFB}
}
