% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice.R
\name{d}
\alias{d}
\title{d}
\usage{
d(x)
}
\arguments{
\item{x}{either a string in the format '\\d*d\\d+' or a vector
with 2 values.}
}
\value{
Either a single vector with length 2
and the names "count" and "face" or a list of those.
It depends on the input and if there is more than 1 die
found.
}
\description{
Coerces an object into a `Dice` object.
}
\details{
If the vector is unnamed, the first value is
interpreted as number of dice and the second is the number
of sides. If there only one number it assumes `Count = 1`.
If the vector has names partial matching is used for
`c("Count", "Faces")`. Please note that partial matching
is case-sensitive.

Do not add white space in strings.

Decimal places are ignored when passing decimal numbers.
}
