% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.qlfactor}
\alias{.qlfactor}
\title{Create an ordered factor for quality levels}
\usage{
.qlfactor(x)
}
\arguments{
\item{x}{An integer vector with values between 0 and 6.}
}
\value{
0 is mapped to "Failed" and the levels 1-6 are
assigned to their according quality levels.
}
\description{
Create an ordered factor for quality levels
}
\details{
This function does not validate its arguments.
}
