% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{dSkillPurged}
\alias{dSkillPurged}
\title{dSkillPurged}
\usage{
dSkillPurged(x, eav, skill, format = c("vector", "df"))
}
\arguments{
\item{x}{A vector of dice sums.}

\item{eav}{A vector with 3 places, each being an effective
attribute value.}

\item{skill}{The skill level}

\item{format}{Determines how the output is generated, one of}
}
\value{
The result depends on the `format` argument.
\describe{
  \item{vector}{This is a vector of probabilities. The names of the
  vector elements denote the sum of dice.}
  \item{df}{A data frame with the columns:
  Outcome (each sum of dice), `p` (probability for this outcome),
  `Remainder` (remaining skill points, with all failed checks being
  projected to a skill point remainder of `-1`.), `QL` (the quality level)}
}
Each vector lists the probabilities for the given sum of 3d20 `x`. </br>
}
\description{
The likelihood for the outcomes of a skill check **excluding** critical
successes and botches.
}
