% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceSum.R
\name{dSumFdN}
\alias{dSumFdN}
\alias{dSum2dN}
\title{dSumFdN}
\usage{
dSumFdN(f = 2L, n = 6L)

dSum2dN(n = 6L)
}
\arguments{
\item{f}{number of dice (integer, `f > 1`)}

\item{n}{number of faces of the dice (integer, `n > 1`)}
}
\value{
A named vector with the probabilities for each sum
of dice.
}
\description{
Calculate the probability of rolling `n` dice with `f` faces.
}
\details{
For two fair n-sided dice (faces 1..n),
the sum distribution is triangular. The functions here have
been tested in a value range of n ∊ [1, 100].

For two fair n-sided dice (faces 1..n),
the sum distribution is triangular.
}
\section{Functions}{
\itemize{
\item \code{dSum2dN()}: `sum2dN` is used for 2 dice because this
method is a low more efficient.

}}
\examples{
dSumFdN() # 2d6
dSumFdN(4, 6) # 4d6
dSumFdN(3, 20) # 3d20
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
