% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFIrstGE.R
\name{findFirstGE}
\alias{findFirstGE}
\title{findFirstGE}
\usage{
findFirstGE(x, target, forward = TRUE)
}
\arguments{
\item{x}{A numeric vector with monotonously increasing values.}

\item{target}{A numeric vector.}

\item{forward}{specifies the search direction. `TRUE` (the default) searches
from left-to-right(low indices to high ones), `FALSE` searches in the other
direction.}
}
\value{
For each target it returns the index of the first `x[i]`
for which `x[i] > target` holds true; if several `x[i]` are equal
it moves the value to the foot of the next slope.

If `target` is greater than any `x`, it returns `NA`.
}
\description{
Find the first index `i` in `x` for which `x[i] >= target` holds true.
}
\details{
This is a helper function for quantile functions.
It may also work outside the value range [0, 1], but has been tested
most extensively within this range.
\itemize{
\item{Leading zeroes}{Leading zeroes are ignored.}
\item{Precision}{The comparison uses a precision of 1E-8. Values with differences
smaller than that will be treated as equal.}
}
}
