% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combat.R
\name{HitPointDistribution}
\alias{HitPointDistribution}
\alias{dhitpoints}
\alias{DiscreteHitPoints}
\alias{phitpoints}
\alias{qhitpoints}
\alias{rhitpoints}
\title{Hit point distribution}
\usage{
dhitpoints(x, eav, w, bl = 20L)

phitpoints(q, eav, w, bl = 20L, lower.tail = TRUE)

qhitpoints(p, eav, w, bl = 20L, lower.tail = TRUE)

rhitpoints(n, eav, w, bl = 20L)
}
\arguments{
\item{x, q}{vector of quantiles, the number of hit points}

\item{eav}{The effective attribute value of the check
(integer, scalar).}

\item{w}{the weapon specifier, a \code{\link[newRoll]{Roll}}
object with die count, faces, and modifier or a string that
can be coerced into it.}

\item{bl}{botch level (integer, [18, 20])}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{X \le x}, otherwise, \eqn{X \ge x}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`,
the length is taken to be the number required.}
}
\value{
`dhitpoints` gives the density, `phitpoints` gives the distribution
function, `qhitpoints` gives the quantile function and
`rhitpoints` generates random deviates.

Throws en error when x is empty or outside
}
\description{
Hit point distribution
}
\examples{
dhitpoints(6, 12, c(Count=2, Dice=6, Mod=0))
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
