% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inc.ordered}
\alias{inc.ordered}
\title{inc.ordered}
\usage{
inc.ordered(x)
}
\arguments{
\item{x}{An ordered factor}
}
\value{
An ordered factor of the same length as `x`.
}
\description{
Increase values of an ordered vector to the next higher level.
At the highest level it doesn't increase further.
}
\keyword{internal}
