% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSFB.R
\name{newCSFB}
\alias{newCSFB}
\alias{is.CSFB}
\alias{print.CSFB}
\alias{as.data.frame.CSFB}
\title{Creates an object to specify the CSFB outcomes of a
TDE check. CSFB are the 4 values of this check
which are *c*ritical successes, regular *s*uccesses,
*f*ails, and *b*otches}
\usage{
newCSFB(cr, su, fa, bo, check = .checkTypes)

is.CSFB(x)

\method{print}{CSFB}(
  x,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = FALSE,
  max = 6L,
  ...
)

\method{as.data.frame}{CSFB}(x, ...)
}
\arguments{
\item{cr, su, fa, bo}{numeric vectors specifying the probabilities
for criticals, successes, fails, and botches.}

\item{check}{The type of check ("Attack", "Parry", "Dodge",
"Skill", or "Attribute").}

\item{x}{object to be coerced to `CSFB`}

\item{digits}{the minimum number of significant digits to be used:
see print.default.}

\item{quote}{logical, indicating whether or not entries should
be printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
right-aligned. The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether
(or what) row names should be printed.}

\item{max}{numeric or NULL, specifying the maximal number of rows
to be printed. By default, when NULL, getOption("max.print") used.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
a `CSFB` object is a list with
the four probability vectors and a string to identify the
type of the check.

`print.CSFB` returns `x` invisibly.

`as.data.frame.CSFB` returns a data frame
}
\description{
Creates an object to specify the CSFB outcomes of a
TDE check. CSFB are the 4 values of this check
which are *c*ritical successes, regular *s*uccesses,
*f*ails, and *b*otches
}
\section{Functions}{
\itemize{
\item \code{is.CSFB()}: S3 method for class 'CSFB'

\item \code{print(CSFB)}: S3 method for class 'CSFB'

\item \code{as.data.frame(CSFB)}: S3 method to coerce object to data frame

}}
\examples{
newCSFB(0.05, 0.45, 0.45, 0.05, "Attack")
print(newCSFB(0.05, 0.45, 0.45, 0.05, "D"))
}
