% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice.R
\name{newDice}
\alias{newDice}
\alias{is.Dice}
\alias{print.Dice}
\title{Creates an object to specify a
combination of dice with count and faces.}
\usage{
newDice(c, f)

is.Dice(x)

\method{print}{Dice}(x, ...)
}
\arguments{
\item{c}{The dice count (integer > 0)}

\item{f}{The faces/sides of the dice (integer > 1)}

\item{x}{a `Dice` object}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
a `Roll` object is a named vector with
count and faces.
}
\description{
Creates an object to specify a
combination of dice with count and faces.
}
\section{Functions}{
\itemize{
\item \code{is.Dice()}: S3 method for class 'Dice'

\item \code{print(Dice)}: S3 method for class 'Dice'

}}
\examples{
newDice(2, 20)
}
