% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice.R
\name{newRoll}
\alias{newRoll}
\alias{is.Roll}
\alias{print.Roll}
\title{Creates an object to specify a die roll
with count, faces, and modifier.}
\usage{
newRoll(c, f, m = 0)

is.Roll(x)

\method{print}{Roll}(x, ...)
}
\arguments{
\item{c}{The dice count (integer > 0)}

\item{f}{The faces/sides of the dice (integer > 1)}

\item{m}{an additive modifier (integer)}

\item{x}{a `Roll` object}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
a `Roll` object is a named vector with
count, faces, and modifier.
}
\description{
Creates an object to specify a die roll
with count, faces, and modifier.
}
\section{Functions}{
\itemize{
\item \code{is.Roll()}: S3 method for class 'Roll'

\item \code{print(Roll)}: S3 method for class 'Roll'

}}
\examples{
newRoll(2, 6, 2)
}
