% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{plotAttributeChecks}
\alias{plotAttributeChecks}
\title{Creates a plot that shows the probabilities of attribute checks.}
\usage{
plotAttributeChecks()
}
\value{
A ggplot object
}
\description{
Creates a plot that shows the probabilities of attribute checks.
}
\examples{
plotAttributeChecks()
}
