% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{plotSkillRemainder}
\alias{plotSkillRemainder}
\title{Skill Check Plots}
\usage{
plotSkillRemainder(eav, skill)
}
\arguments{
\item{eav}{A vector with three integers as effective attribute values.}

\item{skill}{Skill level (integer, scalar)}
}
\value{
A ggplot object
}
\description{
Plot the probabilities of a skill check. Outcomes are plotted on
the x-axis with probabilities on y. Skill point remainders,
}
\examples{
plotSkillRemainder(c(7, 10, 13), 8)
}
