% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{QualityLevels}
\alias{QualityLevels}
\alias{dql}
\alias{pql}
\alias{qql}
\alias{rql}
\title{QualityLevels}
\usage{
dql(x, eav, skill, format = c("vector", "df"))

pql(q, eav, skill, lower.tail = TRUE, format = c("vector", "df"))

qql(p, eav, skill, lower.tail = TRUE, format = c("vector", "df"))

rql(n, eav, skill)
}
\arguments{
\item{x, q}{vector of quantiles, the number of skill points
(integer, 0 <= x <= `r .maxql`).}

\item{eav}{A vector with 3 places, each being an effective
attribute value.}

\item{skill}{The skill value}

\item{format}{Determines how the output is generated, one of}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{X \le x}, otherwise, \eqn{X \ge x}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length
is taken to be the number required.}
}
\value{
The result depends on the `format` argument.
\describe{
  \item{vector}{This is a vector of probabilities. The names of the
  vector elements denote the quality levels.}
  \item{df}{A data frame with the columns:
  Outcome (each sum of dice), `p` (probability for this outcome),
  `Remainder` (remaining skill points), `QL` (the quality level)}
}
}
\description{
Density, distribution function, quantile function and random
generation for the probability distribution
of quality levels of a skill check defined by effective attributes
(`eav`) and  a `skill` rating.
}
\details{
"When making skill checks, the player has a pool of
skill points (SP), equal to the skill rating (SR) of the skill
in question, which can be spent to adjust failed die
rolls into successes. Leftover SP determine the QL" (Ulisses).
A failed skill check is identified by quality level of 0.
}
\examples{
dql(0:6, c(9, 10, 11), 8)
}
\references{
('The Dark Eye' Game reference)[https://tde.ulisses-regelwiki.de/checks.html]
}
