% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectD20.R
\name{rect1d20}
\alias{rect1d20}
\title{rect1d20}
\usage{
rect1d20(eav)
}
\arguments{
\item{eav}{The effective attribute value of the check
(integer, scalar).}
}
\value{
An vector representing the number of events
for each face from 1 to 20. `rect1d20` returns the number
of occurrences.
}
\description{
Create a probability distributions for a d20 truncated at and
rectified to `eav`.
}
\details{
These functions do not validate their arguments.
They are only defined for `1 <= eav <= 20`.
}
