% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectD20.R
\name{RectifiedUniform1d20}
\alias{RectifiedUniform1d20}
\alias{dRect1d20}
\alias{Discrete1d20Uniform}
\alias{pRect1d20}
\alias{qRect1d20}
\alias{rRect1d20}
\title{Rectified 1d20 distribution}
\usage{
dRect1d20(x, eav)

pRect1d20(q, eav, lower.tail = TRUE)

qRect1d20(p, eav, lower.tail = TRUE)

rRect1d20(n, eav)
}
\arguments{
\item{x, q}{vector of quantiles, the sides of the die.}

\item{eav}{The effective attribute value of the check
(integer, scalar).}

\item{lower.tail}{logical; if `TRUE` (default), probabilities are
\eqn{X \le x}, otherwise, \eqn{X \ge x}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`,
the length is taken to be the number required.}
}
\description{
These functions provide information about the
distribution of a d20 which has been truncated and
rectified at the effective attribute value.
`dRect1d20` gives the density, `pRect1d20` gives the
distribution function, `qRect1d20` gives the quantile function,
and `rRect1d20` generates random deviates.
}
\details{
The distribution is a discrete uniform distribution truncated at
and rectified to `eav`. That means all events below ´eav` are
treated to `eav`. Another interpretation is a weighted distribution
with all weights below `eav` being 0 and the weight for `p(eav)`
is `eav`.

This distribution is required to compute the distribution for
skill checks.

`dRect1d20` gives the density, `pRect1d20` gives the
distribution function, `qRect1d20` gives the quantile function,
and `rRect1d20` generates random deviates.
}
\section{Functions}{
\itemize{
\item \code{pRect1d20()}: Distribution function for a truncated and
rectified d20 roll.#'

\item \code{qRect1d20()}: Quantile function for a truncated and
rectified d20 roll.

\item \code{rRect1d20()}: Random generator function for a
truncated and rectified d20 roll.

}}
\seealso{
\link{dSkill}, \link[extraDistr]{ddunif}
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
