% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{SkillPoints}
\alias{SkillPoints}
\alias{dSkillPoints}
\alias{pSkillPoints}
\alias{qSkillPoints}
\alias{rSkillPoints}
\title{SkillPoints}
\usage{
dSkillPoints(x, eav, skill, format = c("vector", "df"))

pSkillPoints(q, eav, skill, lower.tail = TRUE, format = c("vector", "df"))

qSkillPoints(p, eav, skill, lower.tail = TRUE, format = c("vector", "df"))

rSkillPoints(n, eav, skill)
}
\arguments{
\item{x, q}{vector of quantiles, the number of skill points
(integer, 0 <= x <= skill).}

\item{eav}{A vector with 3 places, each being an effective
attribute value.}

\item{skill}{The skill level}

\item{format}{Determines how the output is generated, one of}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{X \le x}, otherwise, \eqn{X \ge x}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length
is taken to be the number required.}
}
\value{
The result depends on the `format` argument.
\describe{
  \item{vector}{This is a vector of probabilities. The names of the
  vector elements denote the remaining skill points.}
  \item{df}{A data frame with the columns:
  Outcome (each sum of dice), `p` (probability for this outcome),
  `Remainder` (remaining skill points), `QL` (the quality level)}
}
}
\description{
Determine the probability of remaining skill points for
a skill check.
}
\details{
The function summarizes all failures under -1 skill points.
}
\examples{
dSkillPoints(0:6, c(9, 10, 11), 8)
}
\references{
\href{https://tde.ulisses-regelwiki.de/checks.html}{'The Dark Eye' Game Reference}
}
