% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceSum.R
\name{dSum2dN}
\alias{dSum2dN}
\title{dSum2dN}
\usage{
dSum2dN(n = 6L)
}
\description{
dSum2dN
}
\details{
For two fair n-sided dice (faces 1..n),
the sum distribution is triangular.
}
\section{Functions}{
\itemize{
\item \code{dSum2dN()}: `sum2dN` is used for 2 dice because this
method is a low more efficient. In fact

}}
