% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEDice-package.R
\docType{package}
\name{TDEDice-package}
\alias{TDEDice}
\alias{TDEDice-package}
\title{TDEDice: Probability Distributions in the World of the Dark Eye}
\description{
Provides functions for the statistics of the role-playing game 'The Dark Eye'.
}
\author{
\strong{Maintainer}: Jan Seifert \email{github@sigurdjanson.anonaddy.com}

Authors:
\itemize{
  \item Jan Seifert \email{github@sigurdjanson.anonaddy.com}
}

}
\keyword{internal}
