% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{binSkillPoints}
\alias{binSkillPoints}
\title{Get the width of bins when binning remaining skill points.}
\usage{
binSkillPoints(sp)
}
\arguments{
\item{sp}{Skill points (scalar, integer)}
}
\value{
A vector starting with the highest quality level.
For the current quality (i.e. the topmost one), it returns the
number of digits above the breakpoint below. Then, since the
lower quality are all fully included, it returns all bin sizes
of the lower bins in order from high to low.
}
\description{
Get the width of bins when binning remaining skill points.
}
