% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{cAttr_BF}
\alias{cAttr_BF}
\alias{cAttr}
\title{Determine the probability of an attribute check}
\usage{
cAttr_BF(eav)

cAttr(eav)
}
\arguments{
\item{eav}{effective attribute value}
}
\value{
A list with 4 vectors: `Critical`, `Success`, `Fail`, `Botch`.
}
\description{
Determine the probability of an attribute check

Determine the probabilities for the outcomes of an attribute check
}
\section{Functions}{
\itemize{
\item \code{cAttr_BF()}: Uses a brute force algorithm. It is
intended for analysis purposes rather than practical use.

}}
\examples{
cAttr(12)
}
