% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{d3D20}
\alias{d3D20}
\title{The density function for rolling a sum of 3d20.}
\usage{
d3D20(x)
}
\arguments{
\item{x}{A vector of dice sums.}
}
\value{
The vector of length 60 starts with indices 1 and 2,
both of which are zero because 3d20 has a sum of at least 3.
}
\description{
The density function for rolling a sum of 3d20.
}
