% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilities.R
\name{dSkill_BF}
\alias{dSkill_BF}
\title{The likelihood for the outcomes of a skill check.}
\usage{
dSkill_BF(eav, skill)
}
\arguments{
\item{s1, s2, s3}{effective attribute}
}
\value{
A vector of length 3*20. Each place represent the
number of events for the sum of dice.
}
\description{
The likelihood for the outcomes of a skill check.
}
\details{
Uses a brute force algorithm. It is intended for analysis purposes
rather than practical use.
}
\examples{
dSkill_BF(7, 10, 13)
}
