% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findFirstGE}
\alias{findFirstGE}
\title{findFirstGE}
\usage{
findFirstGE(x, target, forward = TRUE)
}
\arguments{
\item{x}{A numeric vector with monotonously increasing values.}

\item{target}{A numeric vector.}
}
\value{
For each target it returns the first `x[i]` (from left to right)
for which `x[i] >= target` holds true.
If `target` is greater than any `x`, it returns `NA`.
}
\description{
Find the first index `i` in `x` for which `x[i] >= target` holds true.
}
\details{
This is a helper function for quantile functions.
It also works outside the value range [0, 1], but has been tested
most extensively within this range.

The comparison uses a precision of 1E-8.
}
