% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{qualityLevel}
\alias{qualityLevel}
\title{Determine the quality level based on available skill points.}
\usage{
qualityLevel(sp)
}
\arguments{
\item{sp}{Skill points (integer)}
}
\description{
Determine the quality level based on available skill points.
}
\details{
This function does not validate its arguments.
It's been tested for a value range of [-20, 50].
}
