% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{qualityLevelBreakPoints}
\alias{qualityLevelBreakPoints}
\title{The break points needed to determine the quality level based on
a number of skill points.}
\format{
An object of class \code{integer} of length 5.
}
\usage{
qualityLevelBreakPoints
}
\description{
The break points needed to determine the quality level based on
a number of skill points.
}
\keyword{datasets}
